Rename BumpAlloc -> Arena and simplify
Renaming so that the name reflects the semantics rather than
implementation.
Simplifying internals to reduce its size. Next, I plan on making a
non-owning subclass (which might have the name `Arena`) that can be
passed by value as a means of being a stack-allocator with scope
equivalent to the function's body. This is slightly more succinct than
using the checkpoint or scope commands.
The allocated slabs are now a linked list, and checkpoint has been
made cheaper.